requirements.txt
model_mommy
Django==1.7
django-bootstrap3==5.2.0
django-crispy-forms==1.4.0
django-extensions==1.5.2
django-model-utils==2.2
django-parsley==0.5
MySQL-python==1.2.5
six==1.9.0
